// webapp/src/main/javascript/blomst/src/app/rest.classes.d.ts
export interface AbstractRestEntity extends RestEntity, LinksContainer {
    created: Date;
    modified: Date;
    links: { [index: string]: Link };
}
// ...
export interface UserSelect extends IdentitySelect {
    "@type": "User";
    nickName: string;
    firstName: string;
    lastName: string;
    birthDate: Date;
    language: LanguageForm;
}

export interface Link {
    url: string;
    method: string;
    rel: string;
    description: string;
    type: LinkType;
}

export interface MediaType extends AbstractRestEntity {
    name: string;
    lastModified: Date;
    data: string;
    thumbnail: Thumbnail;
}

export interface EmailForm {
    value: string;
    confirmed: boolean;
}

export interface LanguageForm extends AbstractRestEntity {
    "@type": "Language";
    locale: Locale;
    name: string;
}
// ...
export interface AbstractPostForm extends AbstractLikeableRestEntity {
    "@type": "ImagePost" | "VideoPost" | "TextPost";
    editor: Editor;
    owner: UserSelect;
    source: IdentitySelect;
}
// ...
export interface IdentitySelect extends AbstractRestEntity {
    name: string;
    picture: MediaType;
}
// ...
export interface Editor {
    html: string;
    text: string;
}

export interface AbstractLikeableRestEntity extends AbstractRestEntity {
    views: number;
    likes: boolean;
}